Fixed a bug that happens when trying to buy 0.0001(minimum value) bitcoin

+++ app/src/main/java/com/example/k_trader/base/OrderManager.java
@@ -96,7 +96,7 @@ public class OrderManager {
         return true;
     }
 
-    public JSONObject addOrder(String tag, TradeDataManager.Type type, float units, int price) {
+    public JSONObject addOrder(String tag, TradeDataManager.Type type, double units, int price) {
         Api_Client api = new Api_Client();
         JSONObject result;
         long requestTime = Calendar.getInstance().getTimeInMillis();
